requirements.txt
MySQL-python==1.2.5
PIL==1.1.7
South==0.8.4
amqp==1.4.3
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.14
celery==3.1.8
celery-with-redis==3.0
distribute==0.7.3
django-celery==3.1.1
django-separatedvaluesfield==0.1
fpdf==1.7
ipython==1.1.0
kombu==3.0.12
pytz==2013.9
redis==2.9.1
reportlab==2.7
requests==2.2.1
wsgiref==0.1.2
yolk==0.4.3
mongoengine
hg+http://bitbucket.org/wkornewald/django-nonrel@be48c152abc6b15e45155e2bbcfd69c665ccb536#egg=Django-dev
git+https://github.com/django-nonrel/mongodb-engine@52257b5d90dcfaa564e708264a9cfe591d301fe6#egg=django_mongodb_engine-dev
hg+https://bitbucket.org/wkornewald/djangotoolbox@a8cdf61ba9c0cdc7cbbbd37d693e9a222d9a8e5f#egg=djangotoolbox-dev
django-storages
Wand
